Move itemRenderer to panel.
